+++ src/orego/mcts/McRunnable.java
@@ -27,7 +27,7 @@ public class McRunnable implements Runnable {
 	private McPlayer player;
 
 	/** Number of playouts completed. */
-	private int playoutsCompleted;
+	private long playoutsCompleted;
 
 	/** Array of heuristics. */
 	private HeuristicList heuristics;
@@ -105,7 +105,7 @@ public class McRunnable implements Runnable {
 	}
 
 	/** Returns the number of playouts completed by this runnable. */
-	public int getPlayoutsCompleted() {
+	public long getPlayoutsCompleted() {
 		return playoutsCompleted;
 	}
 
@@ -171,8 +171,8 @@ public class McRunnable implements Runnable {
 	@Override
 	public void run() {
 		boolean limitPlayouts = player.getMillisecondsPerMove() <= 0;
-		int playouts = 0;
-		int limit = player.getPlayoutLimit();
+		long playouts = 0;
+		long limit = player.getPlayoutLimit();
 		while ((limitPlayouts && playouts < limit)
 				|| (!limitPlayouts & getPlayer().shouldKeepRunning())) {
 			performMcRun();
